docs/using-testharness.html

<!DOCTYPE html>  <html> <head>   <title>using-testharness.js</title>   <meta http-equiv="content-type" content="text/html; charset=UTF-8">   <link rel="stylesheet" media="all" href="docco.css" /> </head> <body>   <div id="container">     <div id="background"></div>          <table cellpadding="0" cellspacing="0">       <thead>         <tr>           <th class="docs">             <h1>               using-testharness.js             </h1>           </th>           <th class="code">           </th>         </tr>       </thead>       <tbody>                               <tr id="section-1">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-1">&#182;</a>               </div>               <p><!--</p>             </td>             <td class="code">               <div class="highlight"><pre></pre></div>             </td>           </tr>                               <tr id="section-2">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-2">&#182;</a>               </div>               <p>--></p>

<script src='../js/testharness.js'></script>

<script src='../js/move-log.js'></script>

<iframe id='someIFrame' style='display: none'></iframe>

<script src='../using-testharness.js'></script>

<p><link rel='stylesheet' href='../js/testharness.css'>
This document provides a tutorial for W3C's test framework, known as <code>testharness.js</code>,
which you can download from <a href="https://github.com/jgraham/testharness.js">its GitHub repository</a>.</p>

<p>This tutorial does not assume that you are necessarily familiar with other test frameworks,
but it does expect you to be reasonably proficient with JavaScript (since JavaScript APIs is
what one tests using this).</p>

<p>If you are familiar with other test frameworks such as <a href="http://docs.jquery.com/QUnit">QUnit</a>,
<a href="http://visionmedia.github.com/mocha/">Mocha</a>, or <a href="http://pivotal.github.com/jasmine/">Jasmine</a>
then you should find your way around here relatively easily.</p>

<p>Indeed, <code>testharness.js</code> is not very different from those, though it does have a number of
features that make it particularly well suited to testing APIs implemented by the browser
and exposed <em>to</em> JavaScript rather than those created in JavaScript directly.</p>

<p>This tutorial is actually designed to be runnable. The code you see in the right column
is described on the left, but it is also ran as soon as you load this page. If you scroll
to the bottom, you will see the results of this run.</p>

<p>This documentation was generated using <a href="http://jashkenas.github.com/docco/">Docco</a>, and the
idea behind the way in which it is done is shamelessly stolen from the
<a href="http://pivotal.github.com/jasmine/">Jasmine Documentation</a>.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-3">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-3">&#182;</a>               </div>               <h2>Getting Started</h2>

<p>Let's get started with this code! The first thing you need to do to load <code>testharness.js</code>
is to include it from a <code>script</code> element in the usual way. You can either download your
own copy and set it up locally whichever way you want, or if you're writing a test for
a W3C service you should just point to the W3C copy:</p>

<pre><code>  &lt;script src='/resources/testharness.js'&gt;&lt;/script&gt;
  &lt;script src='/resources/testharnessreport.js'&gt;&lt;/script&gt;
  &lt;link rel='stylesheet' href='/resource/testharness.css'&gt;
</code></pre>

<p>At which point you might rightfully ask why there are two files. The reason for this is
simple: the first one is the actual implementations, and the second one is empty. Why
include an empty file? The idea is that when a specific vendor checks your test suite
out, they can override the content of the <code>testharnessreport.js</code> file. This makes it
possible for them to integrate running the test suite with whatever test reporting tool
they may be using internally.</p>

<p>If you wish the HTML page into which your tests are being run to render results in a nice
and convenient table, you should include an HTML element with ID <code>log</code> where you would
like those results to appear. See the bottom of this document for an example.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-4">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-4">&#182;</a>               </div>               <h2>Basic Testing</h2>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-5">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-5">&#182;</a>               </div>               <p>The most basic usage relies on the <code>test()</code> function, which takes a function with the code
to run, a name of the test, and optionally a literal object providing some additional
options.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-6">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-6">&#182;</a>               </div>               <p>The given function <em>must</em> include at least one assertion; conversely, assertions
can only appear in the context of a test function.
A single test may contain multiple assertions, in which case it is considered to be atomic.
That is to say that a single failed assertion will fail the test, whereas all are required
to pass for the whole test to pass. A document can contain as many tests as you wish it
to.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-7">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-7">&#182;</a>               </div>               <p>The example here contains two assertions, both of which pass.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-8">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-8">&#182;</a>               </div>               <p>But in this example one passes, while the other fails. This causes the entire test to be reported
as a failure.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-9">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-9">&#182;</a>               </div>               <p>In addition to a function and a name, <code>test()</code> can also accept a third parameter being a dictionary
of options. Most of those options are documented in the <a href="#metadata">Including Metadata</a> section
below.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-10">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-10">&#182;</a>               </div>               <p>The only general-purpose option that you can use is <code>timeout</code>. It takes a number of milliseconds
(defaulting to 1000, which the brightest at maths amongst you will have recognised as being equivalent
to one second). If the content of the test takes longer than <code>timeout</code> to run then the test is aborted
and counted as a fail. Since some processing can take longer than one second to run, especially if
you are performing a complex test in a low-end environment (e.g. a basic mobile phone) it can at
times be useful to increase this limit as exemplified here.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-11">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-11">&#182;</a>               </div>               <p>Note that you do not want to use this for asynchronous operations (if only because it won't work).
For those, see the section dedicated to that topic below.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-12">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-12">&#182;</a>               </div>               <h2>Included Assertions</h2>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-13">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-13">&#182;</a>               </div>               <p>There is a good choice of assertions available by default. Most of those take the form
<code>assert_something(actual, expected, description)</code> but several have different signatures.
When a <code>description</code> is part of the signature, it is optional and a string intended for
human consumption in order to provide a better description of which assertion exactly
failed. If you don't provide it, you will get a default error message instead.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-14">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-14">&#182;</a>               </div>               <p><code>assert_true(actual, description)</code> checks that <code>actual</code> is <em>strictly</em> equal to <code>true</code>,
which is to say that it <em>has</em> to be the JavaScript <code>true</code> value and not just someting
that evaluates as "truthy" such as <code>1</code> or <code>"dahut"</code>.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-15">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-15">&#182;</a>               </div>               <p><code>assert_false(actual, description)</code> is the same as <code>assert_true</code> but in reverse. It
has the same strictness about its <code>actual</code> being JavaScript's <code>false</code> and not just
"falsy" (e.g. 0, null)</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-16">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-16">&#182;</a>               </div>               <p><code>assert_equals(actual, expected, description)</code> checks that <code>actual</code> and <code>expected</code>
have the same value (and are in fact the same object). Note that this
comparison is strict and that you should not rely on
whatever automatic type conversions that JavaScript may perform on comparisons.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-17">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-17">&#182;</a>               </div>               <p><code>assert_not_equals(actual, expected, description)</code> is the reverse of <code>assert_equals</code>
and checks that its /actual/ and /expected/ are not the same. The same caveat on
comparison strictness applies, so that values that may seem very similar are still
not equal.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-18">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-18">&#182;</a>               </div>               <p><code>assert_in_array(actual, expected, description)</code> checks that <code>actual</code> is in the array
provided in <code>expected</code>. Any odd member will do, but note that it will not recurse
into the array if it is multidimensional.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-19">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-19">&#182;</a>               </div>               <p><code>assert_array_equals(actual, expected, description)</code> takes an array for both <code>actual</code>
and <code>expected</code>, and validates that they have the same length and that each item is
<code>assert_equals</code> its corresponding member in the other array. Just like the previous
assertion, this is unidimensional.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-20">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-20">&#182;</a>               </div>               <p><code>assert_approx_equals(actual, expected, epsilon, description)</code> takes a numerical <code>actual</code> value
and checks that it is within <code>epsilon</code> of <code>expected</code>. This is notably useful for floating point
calculations in which you know that some drift may occur and you need to check that the outcome
is within a given ballpark &mdash; but it can also be used in other cases.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-21">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-21">&#182;</a>               </div>               <p><code>assert_regexp_match(actual, expected, description)</code> checks that <code>actual</code> matches the <code>expected</code>
regular expression. The latter can be as simple or complex as you wish to make it, and can be
created with flags.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-22">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-22">&#182;</a>               </div>               <p><code>assert_own_property(object, property_name, description)</code> checks that <code>object</code> has a property that
is truly its own (as opposed to inherited down the prototype chain). JavaScripters will recognise
this as checking <code>hasOwnProperty</code>. If you don't know about this important method, you can
<a href="https://developer.mozilla.org/en/JavaScript/Reference/Global_Objects/Object/hasOwnProperty">read up about it on MDN</a>.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-23">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-23">&#182;</a>               </div>               <p><code>assert_inherits(object, property_name, description)</code> complements <code>assert_own_property</code> in that it
similarly checks that the attribute is available on the object, but asserts that it is <em>not</em> the
object's own property but rather has been inherited down the prototype chain.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-24">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-24">&#182;</a>               </div>               <p><code>assert_idl_attribute(object, attribute_name, description)</code> is the same as
<code>assert_inherits</code> and simply aliases it. For clarity, you may be better off sticking
to the previous one.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-25">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-25">&#182;</a>               </div>               <p><code>assert_readonly(object, property_name, description)</code> checks that the given <code>property_name</code>
on <code>object</code> is properly read-only and therefore cannot be set.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-26">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-26">&#182;</a>               </div>               <p><code>assert_throws(code, func, description)</code> is a powerful way of checking that code throws when
and how you expect it to, knowing that the code in <code>func</code> is what must trigger the exception.
This assertion works differently depending on what you pass for <code>code</code>.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-27">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-27">&#182;</a>               </div>               <p>If <code>code</code> is <code>null</code>, then any old exception will do (this is not a particularly recommended check as
the others are more useful).</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-28">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-28">&#182;</a>               </div>               <p>If <code>code</code> is any kind of object, then its <code>name</code> attribute is checked. That attribute must match the
<code>name</code> attribute on the exception being thrown. This means that you can pass a specific <code>DOMException</code>
object here and have it match if it's what is being thrown.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-29">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-29">&#182;</a>               </div>               <p>If <code>code</code> is a string then it must be one of the commonly recognised <code>DOMException</code> names, and it checks
that <code>func</code> throws the corresponding <code>DOMException</code>. For compatibility with older browsers, the old
exception contacts are supported and mapped to the newer name; so for instance you can use
<code>WRONG_DOCUMENT_ERR</code> to mean <code>WrongDocumentError</code>. The latter style is preferred however.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-30">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-30">&#182;</a>               </div>               <!--
`assert_object_equals(actual, expected, description)` checks that two objects are equal by deep-walking
them side by side and making sure that they have the same fields and that those fields have the same
values. It is still considered somewhat experimental.
test(function () {
    assert_object_equals({ foo: "bar" },
                         { foo: "bar" },
                         "Simple objects.");
    assert_object_equals({ top: "here", kids: { list: ["stuff", { leaf: true } ]} },
                         { top: "here", kids: { list: ["stuff", { leaf: true } ]} },
                         "Simple objects.");
}, "Checks object deep-equality");
-->

<p><code>assert_unreached(description)</code> is a very simple assertion the role of which is to
check that some code is indeed unreachable. It only takes a description, and simply
always throws its hands up in disgust whenever it is called. Opposite here you
can see a case in which it is successful (since untouched).</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-31">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-31">&#182;</a>               </div>               <p>Whereas this one fails because the code reaches it.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-32">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-32">&#182;</a>               </div>               <h2>Asynchronous Testing</h2>

<p>It is increasingly rare for Web APIs to be entirely synchronous. Many of the modern ones are very
careful to be asynchronous whenever an operation may take a little time and therefore freeze the
main thread. Testing asynchronous APIs is naturally different from testing synchronous code since
results from asynchronous calls by their very nature do not follow the nicely linear flow of
synchronous calls.</p>

<p>Thankfully, <code>testharness.js</code> makes testing asynchronous APIs pretty much just as easy as testing
synchronous APIs, with all the assertions being the same and only some small details of how the
tests are set up being different. We will start with an example testing that setTimeout works.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-33">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-33">&#182;</a>               </div>               <p>First, instead of calling <code>test(func, name, options)</code> we call <code>async_test(name, options)</code> and
hold on to its return value &mdash; the latter is the object with which we will interact to
control the flow of our asynchronous test. As you can see, the <code>name</code> and <code>options</code> parameters
that <code>async_test()</code> accepts are exactly the same as those used by <code>test()</code>, and <code>options</code> is
just as optional.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-34">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-34">&#182;</a>               </div>               <p>We will use our setTimeout call to perform an assertion, and flag that the test is over. This
will cancel the timeout and if the assertion is successful (in our case it trivially is) then
the test passes. This is performed with two operations: first, the <code>step()</code> method is used to
define the individual test to be run (just as with the first argument to <code>test()</code>); second, the
<code>done()</code> method is called to tell <code>testharness.js</code> that the entire test has run.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-35">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-35">&#182;</a>               </div>               <p>It is often the case that in testing asynchronous code one needs to assign event handlers to
specific <code>onfoo</code> fields of an object. This can be done with <code>step()</code> but it is somewhat cumbersome
since that call to <code>step()</code> needs itself to be wrapped inside a function. There is a shortcut
for precisely this usage: <code>step_func</code>. What it does is take a function exactly like <code>step()</code>
does, but returns a function that can be used directly as an event handler. The XHR example
opposite makes use of that facility.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-36">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-36">&#182;</a>               </div>               <p><a name='metadata'></a></p>

<h2>Including Metadata</h2>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-37">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-37">&#182;</a>               </div>               <p>If you are writing tests for inclusion in the <a href="http://w3c-test.org/framework/">W3C Testing Framework</a>
(and if you're writing those tests for a W3C group, then you really should), then this section can be
of interest to you. If not, you can safely skip it.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-38">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-38">&#182;</a>               </div>               <p>In order to best integrate into the Testing Framework, your tests ought to have metadata. If you have only
one test per HTML file, then your test metadata should be contained in the &lt;head> section of your
document and, again, you can safely skip over this section. If, however, you wish to include an entire
test suite in a single document (which you certainly can do) then it is useful to specify test metadata
for every call to <code>test()</code>.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-39">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-39">&#182;</a>               </div>               <p>This can be achieved very simply by providing the test metadata as part of the third parameter to <code>test()</code>
which we have seen earlier. There are three fields that you can use: <code>help</code> which is a pointer to the
section of the specification that this test is exercising; <code>assert</code> which is an array of assertion description
that your test contains; and <code>author</code> which is simply the author of the test.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-40">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-40">&#182;</a>               </div>               <h2>Advanced Usage</h2>

<p>Most users should not need this section. Read it if you are trying to achieve something
complex that is not working, or if you are curious &mdash; but don't worry if it does not
seem to make much sense as oftentimes you will not need it.</p>

<h3>Complex Setup</h3>

<p>Sometimes it is important either to perform complex setup operations for a test run, or
to modify the overall behaviour of the test run, or both. These can be performed by using
the <code>setup(func, properties)</code>. Either of its arguments is optional, so it can also be called
as <code>setup(func)</code> and <code>setup(properties)</code>.</p>

<p>Once the first test has run, any call to <code>setup()</code> is immediately short-circuited so none of
the examples in this section actually do anyting (you're expected to only have one <code>setup()</code>
anyway) but they are still useful for their illustrative value.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-41">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-41">&#182;</a>               </div>               <p>Essentially, anything that happens in <code>setup()</code>'s function (when used) stays in <code>setup()</code>'s
function. This allows massive failure to happen there and for the test to still attempt a run.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-42">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-42">&#182;</a>               </div>               <p>The <code>properties</code> parameter is a dictionary that can take four fields.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-43">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-43">&#182;</a>               </div>               <p><code>timeout: ms</code>. This sets the timeout not for a single test (as with the other <code>timeout</code>
option) but rather for the entire set of tests in the page. You can use this if you are
concerned about the entire run being slow even though individual tests may not be triggering
their own timeouts.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-44">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-44">&#182;</a>               </div>               <p><code>explicit_done: true | false</code>. Normally, a test run is considered complete (and the report
generated, etc.) whenever the document's load event triggers and all synchronous tests have
reported values (or timed out), and if there are asynchronous tests when those have either
run or timed out as well. In other words, under normal operations the system knows how to
guess when a test run is complete and does not need you to tell it. You may wish to
override this behaviour if you are doing something that could confuse the system, such as
for instance asynchronously loading or generating new tests. If that is the case, then
set <code>explicit_done</code> to <code>true</code> and when you know you're done with all the tests you want
to run, call the global <code>done()</code> function yourself.</p>

<p>It is important to note that if you wish to use <code>explicit_done</code> you <em>must</em> set it to true
<em>before</em> the load event triggers. Otherwise, you will spend a lot of time wondering why
nothing is working in the way you expect.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-45">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-45">&#182;</a>               </div>               <p><code>output_document: Document</code>. By default the runner will log the test results inside an element
with an ID of <code>log</code> inside the same document that the tests are being run in. This works well
for the vast majority of cases, but in some situations you will want something different. The
test run document may be included in a larger one in which you would like to see the test to
appear, or you may wish to open a popup or new tab and write the results there. More typically,
if you are running tests inside an SVG document, providing a <code>log</code> element will do you little
good unless it can be rendered as HTML. In that case, you will want to redirect the output
to an HTML document. For all of those, just pass a Document object to <code>output_document</code>.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-46">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-46">&#182;</a>               </div>               <p><code>explicit_timeout: true | false</code>. In some cases you don't want to handle timeouts yourself
at all (typically because your tests are being run in the context of a larger test runner
that is controlling timeouts on your behalf). If that is the case, then simply set
<code>explicit_timeout</code> to <code>true</code> and have whatever is in charge of controlling timeouts call
the global <code>timeout()</code> function directly.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-47">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-47">&#182;</a>               </div>               <h3>Formatting</h3>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-48">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-48">&#182;</a>               </div>               <p>At times when you wish to report issues in a test suite, or simply log something during development,
it can be very valueable to produce output that is more human-oriented than what <code>toString()</code> is
most likely to provide most of the time (there is only so much you can infer from <code>[object Object]</code>).
For those times, simply use the global <code>format_value(value)</code>. It knows how to format arrays (by recursing
into them), strings with control characters, JavaScript core types including negative zero, and the
more important DOM Node types.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-49">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-49">&#182;</a>               </div>               <h2>Generating Tests</h2>

<p>Writing tests can be a very repetitive endeavour. At times, you simply need to call the same assertion
on a long list of actual and expected values, and when that happens the overhead of the testing boilerplate,
no matter how lightweight it has been made to be, can seem overwhelming.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-50">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-50">&#182;</a>               </div>               <p>In order to make your life ever so simpler, <code>testharness.js</code> provides a very simple function that will call
the same assertion repeatedly on a list of actual and expected values each described by a name. The signature
for that is <code>generate_tests(assert_something, [ [name, actual, expected], ...])</code>.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-51">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-51">&#182;</a>               </div>               <h2>Callbacks</h2>

<p>At times it can be useful to know what is going on inside the test harness so that you can react to it
and build your own behaviour (for instance, producing your own reports or integrating with a larger
testing system that you may be using).</p>

<p>For that purpose, <code>testharness.js</code> provides a set of events that you can be informed of if you so
desire.
The first way in which you can be notified of these events is, if you are in the same context as the
tests being run and can run code before they start, to simply register some callbacks.
The second way to be notified is if you are running the test in its own iframe (or object element)
contained in the document that is your context (any level down) by creating a function with a
specific name that will be called for you. (Be careful with this latter approach that these calls
will not happen across origin boundaries).</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-52">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-52">&#182;</a>               </div>               <p><code>start</code>. Setup has happened and the first test has been created. This is the point at which your
wrapper can do its own setup.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-53">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-53">&#182;</a>               </div>               <p><code>result</code>. Happens whenever a result is produced. It receives a <code>Test</code> object that has a <code>status</code>
field which can be compared to the <code>PASS</code>, <code>FAIL</code>, <code>TIMEOUT</code>, or <code>NOTRUN</code> fields on the same
object, meaning respectively that the test has passed, failed, timed out, or hasn't run at all;
and a <code>message</code> field providing the error message, if any.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-54">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-54">&#182;</a>               </div>               <p><code>complete</code>. Happens when the test run has terminated (successfully or not). It receives an array
of <code>Test</code> objects just like the one passed to <code>result</code> corresponding to all the results, and
a status object describing the state of the entire run, which has a <code>status</code> field which can
be compared to the <code>OK</code>, <code>ERROR</code>, and <code>TIMEOUT</code> fields on the same object and respectively
mean that the suite has entirely succeeded, that it has failed, or that there has been a
time out.</p>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                               <tr id="section-55">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-55">&#182;</a>               </div>               <!--
## Writing Your Own Assertions
-->

<h2>Results</h2>

<p>As promised this is a self-runnable document that includes the results for the test suite
specified by the code in the right column above. You can see the results below:</p>

<p>
 <a href="http://mobiwebapp.eu/"><img alt="MobiWebApp logo" src="http://www.w3.org/2010/09/logomobiwebapp-s.png"></a><br/>
 This project is funded by the European Union through the Seventh Framework Programme (FP7/2010-2012) under grant
 agreement n°257800 - <a href="http://mobiwebapp.eu/">Mobile Web Applications</a> (MobiWebApp).
</p>

<div id='log'></div>             </td>             <td class="code">               <div class="highlight"><pre>undefined</pre></div>             </td>           </tr>                </tbody>     </table>   </div> </body> </html> 
